extract interface actions logic to helper
